reorganise code for better reading flow in output_formatting.rs
